+++ src/test/kotlin/org/wfanet/measurement/api/v2alpha/EventFieldTest.kt
@@ -19,7 +19,7 @@ import kotlin.test.assertFailsWith
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.JUnit4
-import org.wfanet.measurement.api.v2alpha.event_templates.testing.AgeRange
+import org.wfanet.measurement.api.v2alpha.event_templates.testing.TestVideoTemplate.AgeRange
 import org.wfanet.measurement.api.v2alpha.event_templates.testing.TestVideoTemplate
 
 @RunWith(JUnit4::class)
